Move nullable into separate file
