+++ docs/source/technical_details/type_mapping.rst
@@ -4,33 +4,33 @@ Type Mapping
 *OpenAlchemy* translates *OpenAPI* types to *SQLalchemy* types using the
 following mappings:
 
-+--------------+----------------+-----------------+
-| OpenAPI type | OpenAPI format | SQLAlchemy type |
-+==============+================+=================+
-| integer      | *undefined*    | Integer         |
-+--------------+----------------+-----------------+
-|              | int32          | Integer         |
-+--------------+----------------+-----------------+
-|              | int64          | BigInteger      |
-+--------------+----------------+-----------------+
-| number       | *undefined*    | Float           |
-+--------------+----------------+-----------------+
-|              | float          | Float           |
-+--------------+----------------+-----------------+
-| string       | *undefined*    | String          |
-+--------------+----------------+-----------------+
-|              | password       | String          |
-+--------------+----------------+-----------------+
-|              | byte           | String          |
-+--------------+----------------+-----------------+
-|              | binary         | LargeBinary     |
-+--------------+----------------+-----------------+
-|              | date           | Date            |
-+--------------+----------------+-----------------+
-|              | date-time      | DateTime        |
-+--------------+----------------+-----------------+
-| boolean      |                | Boolean         |
-+--------------+----------------+-----------------+
++--------------+----------------+-----------------+-------------------+
+| OpenAPI type | OpenAPI format | SQLAlchemy type | Python type       |
++==============+================+=================+===================+
+| integer      | *undefined*    | Integer         | int               |
++--------------+----------------+-----------------+-------------------+
+|              | int32          | Integer         | int               |
++--------------+----------------+-----------------+-------------------+
+|              | int64          | BigInteger      | int               |
++--------------+----------------+-----------------+-------------------+
+| number       | *undefined*    | Float           | float             |
++--------------+----------------+-----------------+-------------------+
+|              | float          | Float           | float             |
++--------------+----------------+-----------------+-------------------+
+| string       | *undefined*    | String          | str               |
++--------------+----------------+-----------------+-------------------+
+|              | password       | String          | str               |
++--------------+----------------+-----------------+-------------------+
+|              | byte           | String          | str               |
++--------------+----------------+-----------------+-------------------+
+|              | binary         | LargeBinary     | bytes             |
++--------------+----------------+-----------------+-------------------+
+|              | date           | Date            | datetime.date     |
++--------------+----------------+-----------------+-------------------+
+|              | date-time      | DateTime        | datetime.datetime |
++--------------+----------------+-----------------+-------------------+
+| boolean      |                | Boolean         | bool              |
++--------------+----------------+-----------------+-------------------+
 
 String
 ------
