+++ open_alchemy/helpers/__init__.py
@@ -2,6 +2,7 @@
 # pylint: disable=useless-import-alias
 
 from . import peek as peek
+from .calculate_nullable import calculate_nullable as calculate_nullable
 from .define_all import define_all as define_all
 from .get_ext_prop import get_ext_prop as get_ext_prop
 from .merge_all_of import merge_all_of as merge_all_of
