+++ setup.cfg
@@ -2,7 +2,7 @@
 universal = 1
 
 [tool:pytest]
-addopts = -v --cov=open_alchemy --cov=examples/app --cov-config=tests/.coveragerc --flake8 --strict -m "models_file"
+addopts = -v --cov=open_alchemy --cov=examples/app --cov-config=tests/.coveragerc --flake8 --strict
 flakes-ignore =
     *__init__.py ALL
 markers =
